Move clutter_do_later to clutter dir
